refactor: change data type int to long
The result of addString method can overflow out of Integer.MAX_VALUE

The addString method should return long value.
The CalculateViewer class parameter should be passed to long.
The Calculator class should have a local variable as long

+++ src/main/java/calculator/controller/Calculator.java
@@ -27,7 +27,7 @@ public class Calculator {
             case INVALID_PATTERN -> throw new IllegalArgumentException("ERROR: invalid argument");
         }
         String s = userInput.replaceAll("//.\\\\n", "");
-        int sum = stringAdder.addString(s, regex);
+        long sum = stringAdder.addString(s, regex);
 
         calculatorViewer.printResult(sum);
     }
