+++ src/main/java/calculator/model/StringAdder.java
@@ -3,11 +3,11 @@ package calculator.model;
 import java.util.Arrays;
 
 public class StringAdder {
-    public int addString(String input, String delimiter){
+    public long addString(String input, String delimiter){
         String[] numbers = input.split(delimiter);
         return Arrays.stream(numbers)
                 .filter(number->!number.isEmpty())
-                .mapToInt(Integer::parseInt)
+                .mapToLong(Long::parseLong)
                 .sum();
     }
 }
