+++ src/main/java/calculator/viewer/CalculatorViewer.java
@@ -7,7 +7,7 @@ public class CalculatorViewer {
         return userInput;
     }
 
-    public void printResult(int result){
+    public void printResult(long result){
         System.out.println("결과 : " + result);
     }
 }
